requirements.txt
django==5.1.5
djangorestframework==3.15.2 
requests==2.32.3 
celery==5.4.0 
redis==5.2.1 
django-crispy-forms==2.3
django-environ==0.12.0
django-celery-beat==2.7.0